% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_tecan.R
\name{.plate_read_tecan_csv}
\alias{.plate_read_tecan_csv}
\title{Read a Tecan output file as a .csv}
\usage{
.plate_read_tecan_csv(file, temp = FALSE)
}
\arguments{
\item{file}{File path to the file.}

\item{temp}{Whether to include the temperature column}
}
\description{
Read a Tecan output file as a .csv
}
\keyword{internal}
