% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_tecan.R
\name{.plate_read_tecan_xslx}
\alias{.plate_read_tecan_xslx}
\title{Read a tecan output file as a .xlsx}
\usage{
.plate_read_tecan_xslx(file, temp = FALSE)
}
\arguments{
\item{file}{Path to the file}

\item{temp}{Whether to include the temperature column}
}
\description{
Read a tecan output file as a .xlsx
}
\keyword{internal}
