% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{drop_results}
\alias{drop_results}
\title{Drop the lines that are "results" from a biotek file}
\usage{
drop_results(lines)
}
\arguments{
\item{lines}{Character vector of lines. Drops all lines after the 'Results'
headings in biotek .csv files}
}
\description{
Drop the lines that are "results" from a biotek file
}
\keyword{internal}
