% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{get_all_blocks}
\alias{get_all_blocks}
\title{Read all data blocks from a file, with each dataframe being an entry in a
list}
\usage{
get_all_blocks(
  lines,
  temp = FALSE,
  format_well = TRUE,
  include_id = FALSE,
  second_wl = TRUE,
  rownum = FALSE
)
}
\arguments{
\item{lines}{Vector of strings that are lines from a .csv file.}

\item{temp}{Whether to include the temp column of a file.}

\item{format_well}{Whether to format the well ID column of a file.}

\item{include_id}{Whether to include the label / ID information for the data
block as a column in the resulting data frame.}

\item{second_wl}{Whether to include the second possible wavelength in column
names when reading in.}

\item{rownum}{Whether to include row numbers from pre-pivoted blocks when
reading in.}
}
\description{
Read all data blocks from a file, with each dataframe being an entry in a
list
}
\keyword{internal}
