% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{n_rows_from_wells}
\alias{n_rows_from_wells}
\title{Calculate number of rows from given number of wells.}
\usage{
n_rows_from_wells(x)
}
\arguments{
\item{x}{Number of wells in plate, e.g. 384 or 96.}
}
\value{
integer
}
\description{
Calculate number of rows from given number of wells.
}
\keyword{internal}
