% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_metadata.R
\name{plate_add_meta}
\alias{plate_add_meta}
\title{Add \verb{\{plater\{} Formatted Metadata to a Dataset}
\usage{
plate_add_meta(data, file)
}
\arguments{
\item{data}{A tibble::tibble() which contains a 'well' column, that will have
the metadata added to it.}

\item{file}{File patch to a \code{{plater}} formatted \code{.csv} file which contains
the metadat information.}
}
\value{
a tibble::tibble()
}
\description{
Add \verb{\{plater\{} Formatted Metadata to a Dataset
}
\examples{
file_data <- system.file(
  "extdata",
  "20220929_1steptimer20.csv",
  package = "wellr"
)
file_meta <- system.file(
  "extdata",
  "20220929_1steptimer20_metainfo.csv",
  package = "wellr"
)

plate_read_biotek(file_data) \%>\%
  plate_add_meta(file_meta)
}
