% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{plate_read_biotek}
\alias{plate_read_biotek}
\title{Read Biotek Output CSV Files}
\usage{
plate_read_biotek(file, time_average = TRUE, second_wl = FALSE)
}
\arguments{
\item{file}{The filepath to a \code{.csv} file, exported from a Biotek plate
reader.}

\item{time_average}{Logical, whether to average the time points across
different observations (LUM / OD) and pivot them to their own columns.}

\item{second_wl}{Whether to include the second wavelength when reading in
data blocks, or to keep just the label and the first wavelength.}
}
\value{
tibble::tibble()
}
\description{
Read Biotek Output CSV Files
}
\examples{
file_data <- system.file(
  "extdata",
  "2024-02-29_vio_GFP_main.csv",
  package = "wellr"
)

plate_read_biotek(file_data)
plate_read_biotek(file_data, second_wl = TRUE)
}
