% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{plate_read_biotek2}
\alias{plate_read_biotek2}
\title{drop-in replacement for plate_read_biotek}
\usage{
plate_read_biotek2(
  file,
  average_time = TRUE,
  accumulate_time = TRUE,
  interval = 0,
  second_wl = TRUE,
  rename = TRUE
)
}
\description{
drop-in replacement for plate_read_biotek
}
\keyword{internal}
