% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek2.R
\name{plate_read_biotek_wl}
\alias{plate_read_biotek_wl}
\title{Read the \code{wavelength} data blocks from a \emph{biotek} \code{.csv} file.}
\usage{
plate_read_biotek_wl(file, format_well = TRUE)
}
\arguments{
\item{file}{File path to the \code{.csv} file.}

\item{format_well}{Whether to format the \code{well} column of the returned dataframes.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read the \code{wavelength} data blocks from a \emph{biotek} \code{.csv} file.
}
\examples{
file_data <- system.file(
  "extdata",
  "2024-02-29_vio_GFP_main.csv",
  package = "wellr"
)

plate_read_biotek_wl(file_data)
plate_read_biotek_wl(file_data, format = FALSE)
}
