% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{plate_read_meta}
\alias{plate_read_meta}
\title{Read a \code{{plater}} Formatted Metadata File}
\usage{
plate_read_meta(file, sep = ",")
}
\arguments{
\item{file}{File path to a \code{{plater}} formatted \code{.csv} file which contains
metadata information.}

\item{sep}{The separator for the \code{.csv} file, defaults to ",".}
}
\value{
a tibble::tibble()
}
\description{
Read a \code{{plater}} Formatted Metadata File
}
\examples{

file_meta <- system.file(
  "extdata",
  "20220929_1steptimer20_metainfo.csv",
  package = "wellr"
)

plate_read_meta(file_meta)
}
