% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_tecan.R
\name{plate_read_tecan}
\alias{plate_read_tecan}
\title{Read the output of Tecan Plate Readers}
\usage{
plate_read_tecan(file, temp = FALSE)
}
\arguments{
\item{file}{File path to the \code{.xlsx} or \code{.csv} file.}

\item{temp}{Logical, whether to include the temperature column.}
}
\value{
a \code{tibble::tibble()} of the values from the file.
}
\description{
Read the output of Tecan Plate Readers
}
\examples{

# read a .csv tecan file
fl <- system.file(
  'extdata', 'tecanON1.csv',
  package = 'wellr'
)

plate_read_tecan(fl)

# read a .xlsx tecan file
fl <- system.file(
  'extdata', 'tecanON1.xlsx',
  package = 'wellr'
)

plate_read_tecan(fl)

}
