% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{read_data_block}
\alias{read_data_block}
\title{Read a data block from a vector of strings that are the lines of a .csv}
\usage{
read_data_block(lines, temp = FALSE, format_well = TRUE, rownum = FALSE)
}
\arguments{
\item{lines}{The vector of lines for the csv}

\item{temp}{Whether to include the temp column or not.}

\item{format}{Whether to format the well column names.}
}
\description{
Read a data block from a vector of strings that are the lines of a .csv
}
\keyword{internal}
