% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_meta}
\alias{read_meta}
\title{Read a \code{{plater}} formatted metadata file}
\usage{
read_meta(file)
}
\arguments{
\item{file}{to read metadata from.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read a \code{{plater}} formatted metadata file
}
\examples{
file <- system.file("extdata", "plate_layout_96.csv", package = "wellr")
read_meta(file)
}
