% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_check}
\alias{well_check}
\title{Checks if a well ID contains a letter and a number in the correct format.}
\usage{
well_check(x)
}
\arguments{
\item{x}{well ID to check for proper structure.}
}
\value{
A vector the same length as x with a TRUE of FALSE for a a well that
can be read by wellr
}
\description{
Checks if a well ID contains a letter and a number in the correct format.
}
