% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_df_to_mat}
\alias{well_df_to_mat}
\title{Convert a data.frame to a Matrix in Plate Format}
\usage{
well_df_to_mat(df, values_from, well_col = "well", plate = NULL)
}
\arguments{
\item{df}{A dataframe containing at least a column called "well" that
contains well IDs.}

\item{values_from}{Column that contains values to populate matrix.}

\item{well_col}{Column containing the well IDs to extract row and column
information from.}

\item{plate}{Size of the plate, to override the auto-detected plate size.}
}
\value{
A matrix containing the values that were present in the data frame in
the \code{values_from} column
}
\description{
Extracts row and column information from the column specified by \code{well_col}
(defaults to 'well') and uses this to create a matrix that contains values
from the \code{values_from} column.
}
\examples{
library(wellr)

# pivot a long data frame into a matrix based on the row and col columns,
# with values from a given column
plate <- well_plate(8, 12)
plate$value <- seq(nrow(plate))
well_df_to_mat(plate, values_from = "value")

# the function also handles missing values from a plate, filling with with NA
# if certain wells are missing
plate <- well_plate(7, 11)
plate$value <- seq(nrow(plate))
well_df_to_mat(plate, values_from = "value", plate = 384)
}
