% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_df_to_matrix}
\alias{well_df_to_matrix}
\title{Convert a data.frame to a Matrix in Plate Format}
\usage{
well_df_to_matrix(df, values_from, plate = NULL)
}
\arguments{
\item{df}{A dataframe containing at least a column called "well" that
contains well IDs.}

\item{values_from}{Column that contains values to populate matrix.}

\item{plate}{Size of the plate, to override the auto-detected plate size.}
}
\value{

}
\description{
Convert a data.frame to a Matrix in Plate Format
}
\examples{
library(wellr)

# pivot a long data frame into a matrix based on the row and col columns,
# with values from a given column
plate <- well_plate(8, 12)
plate$value <- seq(nrow(plate))
well_df_to_matrix(plate, values_from = "value")

# the function also handles missing values from a plate, filling with with NA
# if certain wells are missing
plate <- well_plate(7, 11)
plate$value <- seq(nrow(plate))
well_df_to_matrix(plate, values_from = "value", plate = 384)
}
