% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_dis}
\alias{well_dis}
\title{Relative Distance Between Given Well and a Reference Row}
\usage{
well_dis(row, col, ref_row, ref_col)
}
\arguments{
\item{row}{Numeric vector of row numbers. Must be equal in length to col.}

\item{col}{Numeric vector of col numbers. Must be equal in length to row.}

\item{ref_row}{Single reference row number.}

\item{ref_col}{Single reference column number.}
}
\value{
a numeric vector.
}
\description{
Relative Distance Between Given Well and a Reference Row
}
\examples{
well_dis(3, 4, ref_row = 5, ref_col = 5)
well_dis(1:8, 1:8, ref_row = well_to_rownum("C4"), ref_col = well_to_colnum("C4"))
}
