% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_dis_plate}
\alias{well_dis_plate}
\title{Relative Distance for an Entire Plate}
\usage{
well_dis_plate(ref_row, ref_col, plate = 96)
}
\arguments{
\item{ref_row}{Row number to calculate distance from.}

\item{ref_col}{Column number to calculate distance from.}

\item{plate}{Plate size to calculate distances for.}
}
\value{
a matrix of relative distances.
}
\description{
Relative Distance for an Entire Plate
}
\examples{

well_dis_plate(5, 5)
}
