% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_format}
\alias{well_format}
\title{Format a Well to Uppercase and Padded Numbers}
\usage{
well_format(x, num_width = 2, uppercase = TRUE)
}
\arguments{
\item{x}{Vector of well IDs to be formatted.}

\item{num_width}{Width to pad out number component with 0's.}

\item{uppercase}{Logical, whether the letter should be upercase.}
}
\value{
Vector of strings as formatted well IDs.
}
\description{
Format a Well to Uppercase and Padded Numbers
}
\examples{
well_format(c("A9", "c3", "h12"))
}
