% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_join}
\alias{well_join}
\title{Join a row and column into a well ID.}
\usage{
well_join(row, col, num_width = 2)
}
\arguments{
\item{row}{either a row letter or number for to be coerced into a letter.}

\item{col}{a column number.}

\item{num_width}{the number of digits to pad out the column number with 0.}
}
\value{
a vector of well IDs as strings.
}
\description{
Joins a column number and a row letter or number into a well ID. i.e. joins
"A" and "1" to "A01" and joins "3 and "10" to "C10".
}
\examples{

well_join(1:3, 4)
well_join(c("A", "B", "H"), 9)
well_join("C", 1:10)
}
