% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_mat_frames_to_df}
\alias{well_mat_frames_to_df}
\title{Turn a Multi-Frame Matrix into a DataFrame}
\usage{
well_mat_frames_to_df(matrix, value_col = "value")
}
\arguments{
\item{matrix}{multi-frame matrix to be converted.}

\item{value_col}{Name to be assigned to the values column.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Takes a matrix where each row is a time point and each column is a well and
converts it to a DataFrame containing a column for the well, a column for the
row, a column for the column, a column for the time and a column for the
values.
}
\examples{
df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
}
