% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_mat_to_df}
\alias{well_mat_to_df}
\title{Convert a Plate Matrix to a DataFrame}
\usage{
well_mat_to_df(matrix, value_col = "value")
}
\arguments{
\item{matrix}{A matrix object to be converted.}

\item{value_col}{The name of the value column in the final DataFrame.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Takes a matrix that is formatted as a plate, and converts it to a long-format
'tidy' data frame with a column for the well ID, the row, the column and the
values that were in the matrix.
}
