% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_plate}
\alias{well_plate}
\title{Create a Tibble of Plate Information}
\usage{
well_plate(nrow = 8, ncol = 12)
}
\arguments{
\item{nrow}{Number of rows to have in the generated plate.}

\item{ncol}{NUmber of columns to have in the generated plate.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Generates a \link[tibble:tibble-package]{tibble} that contains row, col, and well
ID for the size of the plate specified in nrow and ncol. If vectors of length1 are supplied to either nrow or ncol, the contents of the vectors are usedinstead of their numeric value.
}
\examples{
well_plate(nrow = 8, ncol = 12)
well_plate(nrow = 4, ncol = 6)
}
