% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{well_plot}
\alias{well_plot}
\title{Plot a Plate Map}
\usage{
well_plot(data, well, value, colour = "black")
}
\arguments{
\item{data}{Dataframe containing data.}

\item{well}{Name of the column with well IDs.}

\item{value}{Name of the column with the values to plot.}

\item{colour}{Colour of the well borders.}
}
\value{
ggplot2 object.
}
\description{
Plot a Plate Map
}
\examples{
dat <- wellr::well_plate()
dat$value <- rnorm(96)
well_plot(dat, well, value)

}
