% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_reorder_df}
\alias{well_reorder_df}
\title{Title}
\usage{
well_reorder_df(data)
}
\arguments{
\item{data}{Data frame to reorder, than contains a row column and a col column.}
}
\value{

}
\description{
Title
}
