% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_col_num}
\alias{well_to_col_num}
\title{Extract number from well ID.}
\usage{
well_to_col_num(x)
}
\arguments{
\item{x}{Well ID as string in format "A10"}
}
\value{
Numeric verctor of column numbers extracted from well ID.
}
\description{
Extract number from well ID.
}
\examples{
well_to_col_num(c("A10", "c3", "h12"))
well_to_col_num("H08")
well_to_col_num("h8")
}
