% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_colnum}
\alias{well_to_colnum}
\title{Extract number from well ID.}
\usage{
well_to_colnum(x)
}
\arguments{
\item{x}{Well ID as string in format "A10"}
}
\value{

}
\description{
Extract number from well ID.
}
\examples{
well_to_colnum(c("A10", "c3", "h12"))
well_to_colnum("H08")
well_to_colnum("h8")
}
