% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_index}
\alias{well_to_index}
\title{Converts Well ID to a Numeric Index}
\usage{
well_to_index(x, plate = 96, colwise = FALSE)
}
\arguments{
\item{x}{string well ID}

\item{plate}{size of the plate. One of c(6, 12, 24, 96, 384)}

\item{colwise}{if TRUE, index instead down the columns, so H01 is index 8,
A12 is index 89 and B01 is index 2 for a 96 well plate.}
}
\value{
numeric well index.
}
\description{
Indexes along rows first, so A12 is index 12 and B01 is index
13 for a 96 well plate.
}
\examples{

# indexing along the row first
well_to_index(c("A10", "c3", "h12"))
well_to_index("H08")
well_to_index("h8")
well_to_index("C20")

# indexing instead down the column first
well_to_index(c("A10", "c3", "h12"), colwise = TRUE)
well_to_index("H08", colwise = TRUE)
well_to_index("h8", colwise = TRUE)
well_to_index("C20", colwise = TRUE)
}
