% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_row_let}
\alias{well_to_row_let}
\title{Extracts letter from well ID.}
\usage{
well_to_row_let(x)
}
\arguments{
\item{x}{Well ID as string.}
}
\value{
a string which is the letter extracted from a well ID.
}
\description{
Extracts letter from well ID.
}
\examples{
well_to_row_let(c("A10", "c3", "h12"))
well_to_row_let("H08")
well_to_row_let("h8")
well_to_row_let("C20")
}
