% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_row_num}
\alias{well_to_row_num}
\title{Convert Well ID to Row Number}
\usage{
well_to_row_num(x)
}
\arguments{
\item{x}{Well ID as a string.}
}
\value{
Numeric row number.
}
\description{
Calculates the row, but returns it as a number.
}
\examples{
well_to_row_num(c("A10", "c3", "h12"))
well_to_row_num("H08")
well_to_row_num("h8")
well_to_row_num("C20")
}
