% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_rowlet}
\alias{well_to_rowlet}
\title{Extracts letter from well ID.}
\usage{
well_to_rowlet(x)
}
\arguments{
\item{x}{Well ID as string.}
}
\value{
a string which is the letter extracted from a well ID.
}
\description{
Extracts letter from well ID.
}
\examples{
well_to_colnum("A10")
well_to_colnum("H08")
well_to_colnum("h8")
well_to_colnum("C20")
}
