% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_rownum}
\alias{well_to_rownum}
\title{Convert Well ID to Row Number}
\usage{
well_to_rownum(x)
}
\arguments{
\item{x}{Well ID as a string.}
}
\value{
Numeric row number.
}
\description{
Calculates the row, but returns it as a number.
}
\examples{
well_to_colnum("A10")
well_to_colnum("H08")
well_to_colnum("h8")
well_to_colnum("C20")
}
