% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellr-package.R
\docType{package}
\name{wellr-package}
\alias{wellr}
\alias{wellr-package}
\title{wellr: Utilities for Working Plate Based Data}
\description{
{wellr} provides a range of utility functions for working with plate-based data from microtitre plates. Working with column IDs, row numbers and letters and transforming between a matrix and a data frame are all made simpler by {wellr}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rforbiochemists/wellr}
  \item Report bugs at \url{https://github.com/rforbiochemists/wellr/issues}
}

}
\author{
\strong{Maintainer}: Brady Johnston \email{brady.johnston@me.com} (\href{https://orcid.org/0000-0001-6301-2269}{ORCID})

}
\keyword{internal}
