% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{plate_read_biotek}
\alias{plate_read_biotek}
\title{Read Biotek Output CSV Files}
\usage{
plate_read_biotek(file)
}
\arguments{
\item{file}{The filepath to a \code{.csv} file, exported from a Biotek plate
reader.}
}
\value{
tibble::tibble()
}
\description{
Read Biotek Output CSV Files
}
\examples{
file_data <- system.file(
  "extdata",
  "20220929_1steptimer20.csv",
  package = "wellr"
)

plate_read_biotek(file_data)
}
