% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_biotek.R
\name{plate_read_biotek}
\alias{plate_read_biotek}
\title{Read Biotek Output CSV Files}
\usage{
plate_read_biotek(file, time_average = TRUE)
}
\arguments{
\item{file}{The filepath to a \code{.csv} file, exported from a Biotek plate
reader.}

\item{time_average}{Logical, whether to aveage the time points across different observations (LUM / OD) and pivot them mto their own columns.}
}
\value{
tibble::tibble()
}
\description{
Read Biotek Output CSV Files
}
\examples{
file_data <- system.file(
  "extdata",
  "2024-02-29_vio_GFP_main.csv",
  package = "wellr"
)

plate_read_biotek(file_data)
}
