% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_df_to_mat_frames}
\alias{well_df_to_mat_frames}
\title{Convert a DatFrame to a Multi-Frame Matrix}
\usage{
well_df_to_mat_frames(data, values_col, time_col, well_col = "well")
}
\arguments{
\item{data}{A DatFrame to be converted.}

\item{values_col}{The name of the column containing the values.}

\item{time_col}{The name of the column containing the time values.}

\item{well_col}{The name of the column with the well IDs.}
}
\value{
a matrix
}
\description{
Converts a DataFrame to a matrix where each row is a single time point, and
each column is a single well. The wells are in index order, indexing across
the rows so the first values are all from row A.
}
\examples{
df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
}
