% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_format}
\alias{well_format}
\title{Format a Well to Uppercase and Padded Numbers}
\usage{
well_format(x, num_width = 2)
}
\arguments{
\item{x}{Well}

\item{num_width}{}
}
\value{

}
\description{
Format a Well to Uppercase and Padded Numbers
}
\examples{
well_format(c("A9", "c3", "h12"))
}
