% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_from_index}
\alias{well_from_index}
\title{Convert Numeric Index to Well ID.}
\usage{
well_from_index(x, plate = 96, num_width = 2)
}
\arguments{
\item{x}{numeric index to convert to a well ID}

\item{plate}{number of wells in the plate. One of c(6, 12, 24, 96, 384)}

\item{num_width}{number of zeros to pad the column number with to the left.}
}
\value{
a column ID as a vector of strings.
}
\description{
Convert Numeric Index to Well ID.
}
