% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{well_reorder_df}
\alias{well_reorder_df}
\title{Reorder a Plate-Based DataFrame}
\usage{
well_reorder_df(data, well_col = "well", row_col = NULL, col_col = NULL)
}
\arguments{
\item{data}{Data frame to reorder, than contains a row column and a col column.}

\item{well_col}{Column containing the well IDs.}

\item{row_col}{Column containing the row letters or numbers.}

\item{col_col}{Column containing the column numbers.}
}
\value{
A data.frame that has been reordered according to the \code{row} & \code{col}
columns.
}
\description{
Reorder a Plate-Based DataFrame
}
\examples{
df <- well_plate(nrow = 8, ncol = 12)
df <- df[sample(1:96, 96), ]
head(df)

df <- well_reorder_df(df)
head(df)
}
