% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well_to_rownum}
\alias{well_to_rownum}
\title{Convert Well ID to Row Number}
\usage{
well_to_rownum(x)
}
\arguments{
\item{x}{Well ID as a string.}
}
\value{
Numeric row number.
}
\description{
Calculates the row, but returns it as a number.
}
\examples{
well_to_rownum(c("A10", "c3", "h12"))
well_to_rownum("H08")
well_to_rownum("h8")
well_to_rownum("C20")
}
